Assignment-1/requirements.txt
langchain
langchain-openai
python-dotenv
openai